fix(layout): 조각 적합 판정을 방출 높이와 같은 식으로 맞추고 앞 조각 개체의 진단 쪽을 보존한다
PR 리뷰 4건 분석 (#182). 둘은 이 PR이 만든 회귀라 고쳤고, 하나는 부분 유효라 문서·계약
테스트로만 닫았으며, 하나는 반영하지 않았다.

## 고친 것

**조각 적합 판정이 방출 높이와 다른 식을 썼다.** 방출은
`takenHeight - ascentExcess(경계)`인데 적합 루프는 보정 없는 전진량으로 비교했다 —
전진량은 baseline 간격이라 다음 줄의 ascent를 통째로 싣고(100pt 표 줄이면 약 90pt) 그
몫은 경계에서 뒤 조각으로 넘어가는데도. 그래서 키 큰 개체 줄 **바로 앞**의 평범한 줄이
실제로는 들어가는데 거절돼 단이 그 줄만큼 비었다. 실측: 본문 200pt 합성 문단에서 앞
조각이 9줄(147.2pt)에서 끊겼고, 고친 뒤 10줄(163.9pt)을 담는다(조각 높이 합 320.0pt는
불변). 식을 `HwpFragmentLineAdvances.chargedHeight`로 올려 판정과 방출이 한 출처를 쓴다.

**앞 조각과 함께 놓인 미지원 개체의 진단 쪽이 틀렸다.** `inlineControlsPlacedPerFragment`가
서수만 남겨, 문단이 끝난 뒤 도는 `walkUnsupported`가 `cachedPages.count + 1`(마지막
조각의 쪽)로 보고했다. 이 PR이 렌더만 앞 쪽으로 옮기고 진단은 그대로 둬 갈린 회귀다 —
1쪽에 그려진 OLE가 2쪽으로 보고됐다. 서수 → 그 조각의 쪽 맵으로 바꿔 진단이 렌더 쪽을
가리킨다. 값이 없는 서수는 종전 값을 그대로 써 동작이 바이트 단위로 같다.

## 문서·테스트로만 닫은 것

**넓은 단 → 좁은 단 이월에서 조각 높이가 진입 단 폭 기준으로 남는다**는 지적은 기전이
실재하지만 main도 같고(렌더러는 블록 높이로 자르지 않아 글자가 이미 같은 양만큼
넘쳤다) `Sources/HwpKitCore/AGENTS.md` "다단 세부"가 이미 적어 둔 근사다. 개체가 함께
내려가는 것은 오배치가 아니라 자기 글리프를 따라간 결과다. 제안된 처방 둘(개체 배치를
미루기·앵커 클램프)은 개체를 자기 글자에서 떼어 놓아 #164로 되돌아간다. 그래서 근사의
**개체 축**을 문서에 명시하고 그 방향의 계약을 테스트로 고정했다 — 실효 수정(조각
경계·높이를 목적 단 폭으로 재계산)은 별건이고 `FixturePreviewFidelityTests`의 Column
실측 여유가 0.001뿐이라 실측 없이 손댈 자리가 아니다.

## 반영하지 않은 것

**새 속성 키의 DocC Topics 등재.** `HwpAttributedStringKey`는 카탈로그에 등재된 적이
없고(main도 동일), 머지된 커밋 넷이 같은 방식으로 공개 키를 더하면서 카탈로그를 건드리지
않았다. 카탈로그는 "진입점 중심 큐레이션"이라 멤버 수준 링크 전례가 0건이다. 대안인
비공개화는 두 줄 위 형제 `inlineObjectHeight`와 가시성이 갈려 반쪽 네임스페이스가 된다.
등재하려면 타입 전체를 놓고 볼 리포 전역 결정이라 이 PR 범위 밖이다.

테스트 공용 입력·조회(`InlineControlFragmentSupport`)는 세 파일이 함께 쓰므로 자기 파일로
뺐다 (SwiftLint 본문·파일 길이 상한).

블록 스냅샷·렌더 해시 포함 전량 통과. 두 수정 모두 수정 전 코드에서 실패함을 확인했다.
