+++ CHANGELOG.md
@@ -164,7 +164,10 @@
   폭도 목적 단으로 다시 풀어, 좁은 단에서 넓은 단으로 이월된 단 너비 50% 개체가 절반만
   예약한 채 두 배로 그려져 뒤 글자를 덮던 것을 막습니다 — 잇달아 놓인 개체도 각자 자기
   크기로 예약합니다. 각주 번호를 쪽마다 새로 시작하는 문서에서 조각의 참조 번호가 다시
-  쓰여 폭이 달라지면(10) → 9)) 그 뒤의 개체도 함께 옮겨 놓습니다.
+  쓰여 폭이 달라지면(10) → 9)) 그 뒤의 개체도 함께 옮겨 놓습니다. 키 큰 개체가 있는 줄
+  바로 앞의 평범한 줄은 그 개체 줄로 가는 간격이 아니라 자기 몫만 차지하므로, 쪽·단이
+  그 줄만큼 비지 않습니다. 앞 조각과 함께 놓인 미지원 개체의 진단 쪽 번호도 실제로
+  그려진 쪽을 가리킵니다.
 - **Shift+Enter로 나눈 줄 끝에 조판 부호가 그려지던 것을 바로잡았습니다** (#146).
   한 줄 끝 코드 10(U+000A)은 줄 나눔이라 조판 문자열에 남겨야 하는데, 종전에는 그
   글자가 글자 모양의 **영문 글꼴**로 조판돼 그 글꼴이 U+000A에 획 있는 글리프를 가진
