+++ Sources/HwpKitCore/AGENTS.md
@@ -1014,7 +1014,15 @@ paraShape와 같은 값**이어야 한다.
   차트 종류는 미재현 (전부 세로 막대 취급). 자동 제목은 한컴 기본 "차트 제목" 폴백
 - 다단 세부: 캐시 run이 없는 문단의 균형 재배치·조각 높이는 라인 수 기준
   근사이고, 비균등 단으로 이월된 텍스트 조각은 draw 시 그 단 폭으로 다시
-  줄바꿈된다 (블록 프레임은 단 경계 안, 시각적 줄 수는 달라질 수 있음)
+  줄바꿈된다 (블록 프레임은 단 경계 안, 시각적 줄 수는 달라질 수 있음).
+  **넓은 단 → 좁은 단** 이월에서는 줄이 늘어 글자가 블록 프레임 아래로 넘치는데,
+  그 줄에 앵커가 있는 글자처럼 취급 개체도 함께 내려간다 (#164, PR 리뷰) — 개체는
+  자기 글리프가 실제로 그려지는 줄·x에 놓이므로 **글자 대비로는 정확**하고, 넘치는
+  까닭은 조각 높이가 여전히 문단 진입 단 폭 기준 근사라서다(main도 같은 양만큼 글자가
+  넘쳤다 — 렌더러는 블록 높이로 자르지 않는다). 고치려면 조각 경계·높이를 목적 단
+  폭으로 다시 계산해야 한다. **개체를 흐름 폴백으로 되돌리거나 앵커 y를 블록 하단으로
+  클램프하는 처방은 금지** — 개체가 자기 글자에서 떨어져 #164로 되돌아간다. 계약 고정:
+  `testWideToNarrowCarryoverKeepsTheObjectOnItsDrawnMarkerLine`
 - 양쪽 정렬은 draw 시 한글처럼 남는 폭을 공백에만 배분해 재조판한다
   (`Text/HwpWordJustification` — 공백 없는 줄/마지막 줄은 CT 기본).
   측정 (HwpParagraphLayout)은 CT justified 그대로 — 줄바꿈은 동일하다.
